refactor: move advocate interface
